+++ Client/scripts/monster-manual.ts
@@ -61,15 +61,82 @@ class MonsterManual extends Component<MonsterManualState>{
     }
 
     private saveCreature:EventListener = (e:Event) => {
+        e.preventDefault();
         const creature:Partial<Creature> = this.state.activeCreatureIndex ? this.state.creatures[this.state.activeCreatureIndex] : {};
         const form = e.currentTarget as HTMLFormElement;
-        const updatedState = {...this.state};
-        if (this.state.activeCreatureIndex){
-            this.state.creatures[this.state.activeCreatureIndex] = creature as Creature;
+        if (form.checkValidity()){
+
+            const index:HTMLInputElement = form.querySelector(`[name="index"]`);
+            if (index.value.length){
+                creature.index = index.value;
+            }
+
+            // @ts-ignore
+            creature.name = form.querySelector(`[name="name"]`).value;
+            // @ts-ignore
+            creature.size = form.querySelector(`#size`).value;
+            // @ts-ignore
+            creature.alignment = form.querySelector(`#alignment`).value;
+            // @ts-ignore
+            creature.type = form.querySelector(`#type`).value;
+            // @ts-ignore
+            creature.subtype = form.querySelector(`#subtype`).value;
+            // @ts-ignore
+            creature.ac = form.querySelector(`#ac`).value;
+            // @ts-ignore
+            creature.hp = form.querySelector(`#hp`).value;
+            // @ts-ignore
+            creature.hitDice = form.querySelector(`#hitDice`).value;
+            // @ts-ignore
+            creature.speed = form.querySelector(`#speed`).value;
+            // @ts-ignore
+            creature.cr = form.querySelector(`#cr`).value;
+            // @ts-ignore
+            creature.xp = form.querySelector(`#xp`).value;
+            // @ts-ignore
+            creature.str = form.querySelector(`#str`).value;
+            // @ts-ignore
+            creature.dex = form.querySelector(`#dex`).value;
+            // @ts-ignore
+            creature.con = form.querySelector(`#con`).value;
+            // @ts-ignore
+            creature.int = form.querySelector(`#int`).value;
+            // @ts-ignore
+            creature.wis = form.querySelector(`#wis`).value;
+            // @ts-ignore
+            creature.cha = form.querySelector(`#cha`).value;
+
+            // @ts-ignore
+            creature.immunities = form.querySelector(`#immunities`).value;
+            // @ts-ignore
+            creature.resistances = form.querySelector(`#resistances`).value;
+            // @ts-ignore
+            creature.vulnerabilities = form.querySelector(`#vulnerabilities`).value;
+            // @ts-ignore
+            creature.senses = form.querySelector(`#senses`).value;
+            // @ts-ignore
+            creature.languages = form.querySelector(`#languages`).value;
+            // @ts-ignore
+            creature.savingThrows = form.querySelector(`#savingThrows`).value;
+            // @ts-ignore
+            creature.skills = form.querySelector(`#skills`).value;
+
+            // @ts-ignore
+            creature.abilities = form.querySelector(`#abilities`).getData();
+            // @ts-ignore
+            creature.actions = form.querySelector(`#actions`).getData();
+            // @ts-ignore
+            creature.legendaryActions = form.querySelector(`#legendaryActions`).getData();
+
+            const updatedState = {...this.state};
+            if (this.state.activeCreatureIndex){
+                this.state.creatures[this.state.activeCreatureIndex] = creature as Creature;
+            }
+            updatedState.view = "search";
+            updatedState.activeCreatureIndex = null;
+            AddCreature(creature as Creature);
+            this.setState(updatedState);
         }
-        updatedState.view = "search";
-        updatedState.activeCreatureIndex = null;
-        this.setState(updatedState);
     }
 
     private handleNewCreatureButtonClick:EventListener = () => {
@@ -134,7 +201,7 @@ class MonsterManual extends Component<MonsterManualState>{
                 <div class="stats line-normal">
                     <div class="block w-full p-0.5">
                         <h3 class="block font-red font-lg font-bold font-serif">${creature.name}</h3>
-                        <p style="font-style:italic;" class="block font-neutral-900 font-xs">${creature.size}${creature.type}${creature.subtype ? ` ${creature.subtype}` : null}, ${creature.alignment}</p>
+                        <p style="font-style:italic;" class="block font-neutral-900 font-xs">${creature.size}${creature.type ? ` ${creature.type}` : null}${creature.subtype ? ` ${creature.subtype}` : null}, ${creature.alignment}</p>
                     </div>
                     <hr>
                     <p class="block w-full p-0.5 font-red font-sm">
@@ -292,22 +359,27 @@ class MonsterManual extends Component<MonsterManualState>{
                 this.setAttribute("state", "editing");
                 const creature = this.state.creatures?.[this.state.activeCreatureIndex] ?? null;
                 view = html`
-                    <form @submit=${this.saveCreature} class="w-full p-0.5" grid="columns 1 gap-0.5">
+                    <form @submit=${this.saveCreature} class="w-full p-0.5" grid="columns 1 gap-1">
                         <input type="hidden" name="index" value="${creature?.index ?? null}">
                         <div class="w-full header" flex="row nowrap items-center">
-                            <button role="button" @click=${this.backToSearch} aria-label="Back" tooltip class="button -icon-only -text -grey -round mr-0.5">
+                            <button type="button" @click=${this.backToSearch} aria-label="Back" tooltip class="button -icon-only -text -grey -round mr-0.5">
                                 <i>
                                     <svg aria-hidden="true" focusable="false" role="img" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 448 512"><path fill="currentColor" d="M231.536 475.535l7.071-7.07c4.686-4.686 4.686-12.284 0-16.971L60.113 273H436c6.627 0 12-5.373 12-12v-10c0-6.627-5.373-12-12-12H60.113L238.607 60.506c4.686-4.686 4.686-12.284 0-16.971l-7.071-7.07c-4.686-4.686-12.284-4.686-16.97 0L3.515 247.515c-4.686 4.686-4.686 12.284 0 16.971l211.051 211.05c4.686 4.686 12.284 4.686 16.97-.001z"></path></svg>
                                 </i>
                             </button>
                             <div class="input w-full" style="flex:1;">
-                                <input type="text"  required .value=${creature?.name ?? null} .placeholder=${!creature?.name ? "Untitled Creature" : null}>
+                                <input type="text" name="name" required .value=${creature?.name ?? null} .placeholder=${!creature?.name ? "Untitled Creature" : null}>
                             </div>
+                            <button type="submit" aria-label="Save" tooltip class="button -icon-only -text -grey -round ml-0.5">
+                                <i>
+                                    <svg aria-hidden="true" focusable="false" role="img" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 448 512"><path fill="currentColor" d="M433.941 129.941l-83.882-83.882A48 48 0 0 0 316.118 32H48C21.49 32 0 53.49 0 80v352c0 26.51 21.49 48 48 48h352c26.51 0 48-21.49 48-48V163.882a48 48 0 0 0-14.059-33.941zM288 64v96H96V64h192zm128 368c0 8.822-7.178 16-16 16H48c-8.822 0-16-7.178-16-16V80c0-8.822 7.178-16 16-16h16v104c0 13.255 10.745 24 24 24h208c13.255 0 24-10.745 24-24V64.491a15.888 15.888 0 0 1 7.432 4.195l83.882 83.882A15.895 15.895 0 0 1 416 163.882V432zM224 232c-48.523 0-88 39.477-88 88s39.477 88 88 88 88-39.477 88-88-39.477-88-88-88zm0 144c-30.879 0-56-25.121-56-56s25.121-56 56-56 56 25.121 56 56-25.121 56-56 56z"></path></svg>
+                                </i>
+                            </button>
                         </div>
-                        <div grid="columns 2 gap-0.5">
+                        <div grid="columns 2 gap-1">
                             <div class="select">
                                 <label for="size">Size</label>
-                                <select name="size" id="size" required>
+                                <select id="size" required>
                                     <option .selected=${creature?.size === "Tiny"}>Tiny</option>
                                     <option .selected=${creature?.size === "Small"}>Small</option>
                                     <option .selected=${creature?.size === "Medium"}>Medium</option>
@@ -318,7 +390,7 @@ class MonsterManual extends Component<MonsterManualState>{
                             </div>
                             <div class="select">
                                 <label for="alignment">Alignment</label>
-                                <select name="alignment" id="alignment" required>
+                                <select id="alignment" required>
                                     <option value="unaligned">Unaligned</option>
                                     <option .selected=${creature?.alignment === "any alignment"}>Any Alignment</option>
                                     <option .selected=${creature?.alignment === "lawful good"}>Lawful good</option>
@@ -340,7 +412,7 @@ class MonsterManual extends Component<MonsterManualState>{
                             </div>
                             <div class="input">
                                 <label for="type">Type</label>
-                                <input name="type" id="type" list="types" required .value=${creature?.type ?? null}>
+                                <input id="type" list="types" required .value=${creature?.type ?? null}>
                                 <datalist id="types">
                                     <option value="aberration"></option>
                                     <option value="beast"></option>
@@ -361,7 +433,7 @@ class MonsterManual extends Component<MonsterManualState>{
                             </div>
                             <div class="input">
                                 <label for="subtype">Subtype</label>
-                                <input name="subtype" id="subtype" list="subtypes" .value=${creature?.subtype ?? null}>
+                                <input id="subtype" list="subtypes" .value=${creature?.subtype ?? null}>
                                 <datalist id="subtypes">
                                     <option value="any race"></option>
                                     <option value="demon"></option>
@@ -411,7 +483,7 @@ class MonsterManual extends Component<MonsterManualState>{
                                 <input type="number" id="xp" min="0" step="1" value="${creature?.xp ?? 0}" required>
                             </div>
                         </div>
-                        <div grid="columns 3 gap-0.5">
+                        <div grid="columns 3 gap-1">
                             <div class="input">
                                 <label for="str">Strength</label>
                                 <input type="number" id="str" min="1" max="30" step="1" value="${creature?.str ?? 10}" required>
@@ -465,6 +537,63 @@ class MonsterManual extends Component<MonsterManualState>{
                             <label for="skills">Skills</label>
                             <input type="text" id="skills" .value="${creature?.skills?.replace("—", "") ?? null}">
                         </div>
+                        <creature-info-table id="abilities">
+                            <h4 class="block w-full font-medium font-sm font-grey-800 pl-0.125">Abilities</h4>
+                            ${creature?.abilities ? html`
+                                ${creature.abilities.map(ability => {
+                                    return html`
+                                        <table-row>
+                                            <row-title>
+                                                <input required value="${ability.name}" placeholder="Name">
+                                                <button class="delete" type="button" aria-label="Remove ability" tooltip>
+                                                    <svg aria-hidden="true" focusable="false" role="img" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 448 512"><path fill="currentColor" d="M296 432h16a8 8 0 0 0 8-8V152a8 8 0 0 0-8-8h-16a8 8 0 0 0-8 8v272a8 8 0 0 0 8 8zm-160 0h16a8 8 0 0 0 8-8V152a8 8 0 0 0-8-8h-16a8 8 0 0 0-8 8v272a8 8 0 0 0 8 8zM440 64H336l-33.6-44.8A48 48 0 0 0 264 0h-80a48 48 0 0 0-38.4 19.2L112 64H8a8 8 0 0 0-8 8v16a8 8 0 0 0 8 8h24v368a48 48 0 0 0 48 48h288a48 48 0 0 0 48-48V96h24a8 8 0 0 0 8-8V72a8 8 0 0 0-8-8zM171.2 38.4A16.1 16.1 0 0 1 184 32h80a16.1 16.1 0 0 1 12.8 6.4L296 64H152zM384 464a16 16 0 0 1-16 16H80a16 16 0 0 1-16-16V96h320zm-168-32h16a8 8 0 0 0 8-8V152a8 8 0 0 0-8-8h-16a8 8 0 0 0-8 8v272a8 8 0 0 0 8 8z"></path></svg>
+                                                </button>
+                                            </row-title>
+                                            <textarea required placeholder="Description" rows="4">${ability.desc}</textarea>
+                                        </table-row>
+                                    `;
+                                })}
+                            ` : null}
+                            <button type="button" class="add">Add Ability</button>
+                        </creature-info-table>
+                        <creature-info-table id="actions">
+                            <h4 class="block w-full font-medium font-sm font-grey-800 pl-0.125">Actions</h4>
+                            ${creature?.actions ? html`
+                                ${creature.actions.map(action => {
+                                    return html`
+                                        <table-row>
+                                            <row-title>
+                                                <input required value="${action.name}" placeholder="Name">
+                                                <button class="delete" type="button" aria-label="Remove action" tooltip>
+                                                    <svg aria-hidden="true" focusable="false" role="img" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 448 512"><path fill="currentColor" d="M296 432h16a8 8 0 0 0 8-8V152a8 8 0 0 0-8-8h-16a8 8 0 0 0-8 8v272a8 8 0 0 0 8 8zm-160 0h16a8 8 0 0 0 8-8V152a8 8 0 0 0-8-8h-16a8 8 0 0 0-8 8v272a8 8 0 0 0 8 8zM440 64H336l-33.6-44.8A48 48 0 0 0 264 0h-80a48 48 0 0 0-38.4 19.2L112 64H8a8 8 0 0 0-8 8v16a8 8 0 0 0 8 8h24v368a48 48 0 0 0 48 48h288a48 48 0 0 0 48-48V96h24a8 8 0 0 0 8-8V72a8 8 0 0 0-8-8zM171.2 38.4A16.1 16.1 0 0 1 184 32h80a16.1 16.1 0 0 1 12.8 6.4L296 64H152zM384 464a16 16 0 0 1-16 16H80a16 16 0 0 1-16-16V96h320zm-168-32h16a8 8 0 0 0 8-8V152a8 8 0 0 0-8-8h-16a8 8 0 0 0-8 8v272a8 8 0 0 0 8 8z"></path></svg>
+                                                </button>
+                                            </row-title>
+                                            <textarea required placeholder="Description" rows="4">${action.desc}</textarea>
+                                        </table-row>
+                                    `;
+                                })}
+                            ` : null}
+                            <button type="button" class="add">Add Action</button>
+                        </creature-info-table>
+                        <creature-info-table id="legendaryActions">
+                            <h4 class="block w-full font-medium font-sm font-grey-800 pl-0.125">Legendary Actions</h4>
+                            ${creature?.legendaryActions ? html`
+                                ${creature.legendaryActions.map(action => {
+                                    return html`
+                                        <table-row>
+                                            <row-title>
+                                                <input required value="${action.name}" placeholder="Name">
+                                                <button class="delete" type="button" aria-label="Remove legendary action" tooltip>
+                                                    <svg aria-hidden="true" focusable="false" role="img" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 448 512"><path fill="currentColor" d="M296 432h16a8 8 0 0 0 8-8V152a8 8 0 0 0-8-8h-16a8 8 0 0 0-8 8v272a8 8 0 0 0 8 8zm-160 0h16a8 8 0 0 0 8-8V152a8 8 0 0 0-8-8h-16a8 8 0 0 0-8 8v272a8 8 0 0 0 8 8zM440 64H336l-33.6-44.8A48 48 0 0 0 264 0h-80a48 48 0 0 0-38.4 19.2L112 64H8a8 8 0 0 0-8 8v16a8 8 0 0 0 8 8h24v368a48 48 0 0 0 48 48h288a48 48 0 0 0 48-48V96h24a8 8 0 0 0 8-8V72a8 8 0 0 0-8-8zM171.2 38.4A16.1 16.1 0 0 1 184 32h80a16.1 16.1 0 0 1 12.8 6.4L296 64H152zM384 464a16 16 0 0 1-16 16H80a16 16 0 0 1-16-16V96h320zm-168-32h16a8 8 0 0 0 8-8V152a8 8 0 0 0-8-8h-16a8 8 0 0 0-8 8v272a8 8 0 0 0 8 8z"></path></svg>
+                                                </button>
+                                            </row-title>
+                                            <textarea required placeholder="Description" rows="4">${action.desc}</textarea>
+                                        </table-row>
+                                    `;
+                                })}
+                            ` : null}
+                            <button type="button" class="add">Add Legendary Action</button>
+                        </creature-info-table>
                     </form>
                 `;
                 break;
@@ -494,3 +623,51 @@ class MonsterManual extends Component<MonsterManualState>{
     }
 }
 customElements.define("monster-manual", MonsterManual);
+
+class CreatureInfoTable extends HTMLElement{
+    private addRowButton:HTMLButtonElement;
+    constructor(){
+        super();
+        this.addRowButton = this.querySelector("button.add");
+    }
+    private addRow:EventListener = () => {
+        const row = document.createElement("table-row");
+        row.innerHTML = `
+            <row-title>
+                <input required value="" placeholder="Name">
+                <button class="delete" type="button" aria-label="Remove legendary action" tooltip>
+                    <svg aria-hidden="true" focusable="false" role="img" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 448 512"><path fill="currentColor" d="M296 432h16a8 8 0 0 0 8-8V152a8 8 0 0 0-8-8h-16a8 8 0 0 0-8 8v272a8 8 0 0 0 8 8zm-160 0h16a8 8 0 0 0 8-8V152a8 8 0 0 0-8-8h-16a8 8 0 0 0-8 8v272a8 8 0 0 0 8 8zM440 64H336l-33.6-44.8A48 48 0 0 0 264 0h-80a48 48 0 0 0-38.4 19.2L112 64H8a8 8 0 0 0-8 8v16a8 8 0 0 0 8 8h24v368a48 48 0 0 0 48 48h288a48 48 0 0 0 48-48V96h24a8 8 0 0 0 8-8V72a8 8 0 0 0-8-8zM171.2 38.4A16.1 16.1 0 0 1 184 32h80a16.1 16.1 0 0 1 12.8 6.4L296 64H152zM384 464a16 16 0 0 1-16 16H80a16 16 0 0 1-16-16V96h320zm-168-32h16a8 8 0 0 0 8-8V152a8 8 0 0 0-8-8h-16a8 8 0 0 0-8 8v272a8 8 0 0 0 8 8z"></path></svg>
+                </button>
+            </row-title>
+            <textarea required placeholder="Description" rows="4"></textarea>
+        `;
+        this.insertBefore(row, this.addRowButton);
+        const deleteButton = row.querySelector("button.delete");
+        deleteButton.addEventListener("click", this.deleteRow);
+    }
+    private deleteRow:EventListener = (e:Event) => {
+        const button = e.currentTarget as HTMLElement;
+        button.parentElement.parentElement.remove();
+    }
+    public getData(){
+        const output = [];
+        const rows = Array.from(this.querySelectorAll("table-row"));
+        for (let i = 0; i < rows.length; i++){
+            const title = rows[i].querySelector("input");
+            const description = rows[i].querySelector("textarea");
+            output.push({
+                name: title.value,
+                desc: description.value
+            });
+        }
+        return output;
+    }
+    connectedCallback(){
+        this.addRowButton.addEventListener("click", this.addRow);
+        const removeButton = Array.from(this.querySelectorAll("button.delete"));
+        for (let i = 0; i < removeButton.length; i++){
+            removeButton[i].addEventListener("click", this.deleteRow);
+        }
+    }
+}
+customElements.define("creature-info-table", CreatureInfoTable);